+++ lib/rubocop/cop/rspec_structure_cops.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "rspec_structure/condition_in_example"